requirements.txt
pandas
numpy
matplotlib
seaborn
psycopg2-binary
sqlalchemy
yfinance
alpha-vantage
flask
flask-cors
scikit-learn
tensorflow
python-dotenv
joblib
xgboost
requests
python-dateutil
imbalanced-learn
